cleanup

+++ Engine/Engine/Code/Framework.cpp
@@ -160,7 +160,7 @@ LRESULT CALLBACK WndProc(HWND hwnd, UINT msg, WPARAM wParam, LPARAM Lparam)
 // returns a randomized float
 float Framework::GetRandomFloat(float min, float max)
 {
-	unsigned int seed = std::chrono::system_clock::now().time_since_epoch().count();
+	unsigned int seed = (unsigned int)std::chrono::system_clock::now().time_since_epoch().count();
 	std::mt19937 generator(seed);
 	std::uniform_real_distribution<float> distribution(min, max);
